minor UI fix
