+++ src/components/common/index.ts
@@ -51,4 +51,3 @@ export * from './helpers/compareVersion'
 export * from './header/PageHeader'
 export * from './error/ErrorBar'
 export * from './DetectBottom'
-export * from './formFields/ResizableTextArea'
