+++ src/components/userGroups/Group.tsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect, useContext, useRef } from 'react';
-import { showError, Progressing, DeleteDialog, deepEqual, ResizableTextarea } from '../common';
+import { showError, Progressing, DeleteDialog, deepEqual } from '../common';
+import { ResizableTextarea } from '../configMaps/ConfigMap';
 import { saveGroup, deleteGroup } from './userGroup.service';
 
 import {
@@ -8,6 +9,7 @@ import {
     EntityTypes,
     ActionTypes,
     CreateGroup,
+    K8sPermissionFilter,
 } from './userGroups.types';
 import './UserGroup.scss';
 import { toast } from 'react-toastify';
@@ -117,7 +119,7 @@ export default function GroupForm({
                         entity: EntityTypes.CLUSTER,
                         action: permission.action.value,
                         cluster: permission.cluster.label,
-                        group: permission.group.value === '*' ? '' : permission.group.value,
+                        group: permission.group.value === '*' ? '' : permission.group.value, 
                         kind: permission.kind.value === '*' ? '' : permission.kind.label,
                         namespace: permission.namespace.value === '*' ? '' : permission.namespace.value,
                         resource: permission.resource.find((entity) => entity.value === '*')
